+++ src/apps/geoserver/wfs/src/test/java/org/geoserver/cloud/wfs/app/WfsApplicationTest.java
@@ -5,35 +5,18 @@
 package org.geoserver.cloud.wfs.app;
 
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.io.TempDir;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.DynamicPropertyRegistry;
-import org.springframework.test.context.DynamicPropertySource;
 import org.xmlunit.assertj3.XmlAssert;
 
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.util.Map;
 
 @SpringBootTest(classes = WfsApplication.class, webEnvironment = WebEnvironment.RANDOM_PORT)
-@ActiveProfiles("test")
-class WfsApplicationTest {
+abstract class WfsApplicationTest {
 
-    static @TempDir Path tmpdir;
-    static Path datadir;
-
-    @DynamicPropertySource
-    static void setUpDataDir(DynamicPropertyRegistry registry) throws IOException {
-        datadir = Files.createDirectory(tmpdir.resolve("datadir"));
-        registry.add("geoserver.backend.data-directory.location", datadir::toAbsolutePath);
-    }
-
-    private TestRestTemplate restTemplate = new TestRestTemplate("admin", "geoserver");
+    protected TestRestTemplate restTemplate = new TestRestTemplate("admin", "geoserver");
 
     @Test
     void owsGetCapabilitiesSmokeTest(@LocalServerPort int servicePort) {
