+++ src/apps/geoserver/wfs/src/test/resources/bootstrap-datadir.yml
